Break up the cli commands from the main server directory
