+++ cmd/root.go
@@ -1,8 +1,7 @@
-package server
+package cmd
 
 import (
 	"fmt"
-	"os"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -12,7 +11,7 @@ var RootCmd = &cobra.Command{
 	Use:   "cyboard",
 	Short: "Scoring Engine",
 	Long:  `This will start the Scoring Engine`,
-	Run:   rootRun,
+	Run:   RootRun,
 }
 
 var CfgFile string
@@ -20,6 +19,8 @@ var CfgFile string
 func init() {
 	cobra.OnInitialize(initConfig)
 	RootCmd.PersistentFlags().StringVar(&CfgFile, "config", "", "config file (default is $HOME/.cyboard/config.toml)")
+	RootCmd.AddCommand(ServerCmd)
+	RootCmd.AddCommand(CheckCmd)
 }
 
 func initConfig() {
@@ -35,7 +36,8 @@ func initConfig() {
 	}
 }
 
-func rootRun(cmd *cobra.Command, args []string) {
+func RootRun(cmd *cobra.Command, args []string) {
+	// For debugging purposes
 	fmt.Println(viper.GetString("appname"))
 	fmt.Println(viper.GetString("server.ip"))
 	fmt.Println(viper.GetString("server.http_port"))
@@ -46,19 +48,5 @@ func rootRun(cmd *cobra.Command, args []string) {
 	err := cmd.Help()
 	if err != nil {
 		Logger.Println(err)
-	}
-}
-
-func addCommands() {
-	RootCmd.AddCommand(ServerCmd)
-	RootCmd.AddCommand(CheckCmd)
-}
-
-func Execute() {
-	addCommands()
-	err := RootCmd.Execute()
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(-1)
 	}
 }
