+++ server/mongodb.go
@@ -13,11 +13,6 @@ import (
 
 var mongodbSession *mgo.Session
 
-func init() {
-	ServerCmd.PersistentFlags().String("mongodb_uri", "mongodb://127.0.0.1", "Address of MongoDB in use")
-	viper.BindPFlag("database.mongodb_uri", ServerCmd.PersistentFlags().Lookup("mongodb_uri"))
-}
-
 func DBSession() *mgo.Session {
 	if mongodbSession == nil {
 		log.Println("Making new MongoDB Session")
