+++ server/run.go
@@ -7,25 +7,13 @@ import (
 	"os"
 
 	"github.com/codegangsta/negroni"
-	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
-var ServerCmd = &cobra.Command{
-	Use:   "server",
-	Short: "Web server for static pages and api",
-	Long:  `This will run the web server`,
-	Run:   serverRun,
-}
+// Logger is used to send logging messages to stdout.
+var Logger = log.New(os.Stdout, " ", log.Ldate|log.Ltime|log.Lshortfile)
 
-func init() {
-	ServerCmd.Flags().Int("http_port", 8080, "HTTP Port for cyboard used for redirect")
-	ServerCmd.Flags().Int("https_port", 1443, "HTTPS Port for cyboard")
-	viper.BindPFlag("server.http_port", ServerCmd.Flags().Lookup("http_port"))
-	viper.BindPFlag("server.https_port", ServerCmd.Flags().Lookup("https_port"))
-}
-
-func serverRun(cmd *cobra.Command, args []string) {
+func Run() {
 	// MongoDB setup
 	CreateUniqueIndexes()
 	// Web Server Setup
